+++ messaging/WebClient.m
@@ -374,9 +374,9 @@ static WebClient *instance = nil;
 //    
 //}
 
--(void)getPostWithRemoteKey:(int)remoteKey withCallbackBlock:(void (^) (BOOL success, GLPPost *post))callbackBlock
+-(void)getPostWithRemoteKey:(NSInteger)remoteKey withCallbackBlock:(void (^) (BOOL success, GLPPost *post))callbackBlock
 {
-    NSString *path = [NSString stringWithFormat:@"posts/%d/", remoteKey];
+    NSString *path = [NSString stringWithFormat:@"posts/%d/", (int)remoteKey];
     
     
     [self getPath:path parameters:self.sessionManager.authParameters success:^(AFHTTPRequestOperation *operation, id responseObject) {
